+++ src/world/Units/Creatures/AIInterface.h
@@ -183,7 +183,7 @@ struct AI_Spell
 {
     ~AI_Spell() { autocast_type = (uint32) - 1; }
     uint32 entryId;
-    uint32 instance_mode;
+    uint8 instance_mode;
     uint16 agent;
     uint32 procChance;
     SpellInfo* spell;
